test: uncomment and fix runner tests to run on montgomery domain
As now we can run operations on Montgomery domain for single-precision
operands, tests commented out in the first commit are now reverted back.
Additionally, as EC runner test turned out to be using wrong values for inputs,
the inputs and expected outcome have been corrected.

The two runner tests in EC dialect are separated and the timeout limitation
has changed to "moderate" to prevent a timeout failure.