fix: patch nvptx lowering
Actual root reason of DAG->DAG size mismatch was that the
`ComputePTXValueVTs` only accounted for i128. I just add the same logic
for i256 and the size mismatch is resolved. With this, the kernel
outlining patch is not really necessary now but I believe it's still
good to have.

Failed assertion:
https://github.com/llvm/llvm-project/blob/5ed852f/llvm/lib/Target/NVPTX/NVPTXISelLowering.cpp#L3381

resolves #75