test(field): check interesting op counts for extension field lowering strictly
In the extension field-to-field lowering test, the counts of multiplication,
square, and inversion operations are checked more strictly. The checks are set
to be as strict as possible as the next commit will improve multiplication counts.

Additionally, counting for `mod_arith.cmp` has been added to the `field.cmp`
lowering test.

The **inversion lowering test** target operation was simplified from `a * a^(-1)`
to `a^(-1)` for clearer verification.